Move all the special logic out into kind files
