+++ src/components/clusterexplorer/node.resource.ts
@@ -4,12 +4,10 @@ import { Kubectl } from '../../kubectl';
 import * as kubectlUtils from '../../kubectlUtils';
 import { Host } from '../../host';
 import * as kuberesources from '../../kuberesources';
-import { Pod, ObjectMeta } from '../../kuberesources.objectmodel';
+import { ObjectMeta } from '../../kuberesources.objectmodel';
 import { kubefsUri } from '../../kuberesources.virtualfs';
 import { ClusterExplorerNode, ClusterExplorerNodeImpl, ClusterExplorerResourceNode } from './node';
-import { MessageNode } from './node.message';
-import { resourceNodeCreate, getChildSources, getUICustomiser } from './resourcenodefactory';
-import { ConfigurationValueNode } from './node.configurationvalue';
+import { getChildSources, getUICustomiser } from './resourcenodefactory';
 
 export abstract class ResourceNode extends ClusterExplorerNodeImpl implements ClusterExplorerResourceNode {
     readonly kindName: string;
@@ -60,59 +58,6 @@ export interface ResourceExtraInfo {
     readonly labelSelector?: any;
     readonly namespaceInfo?: kubectlUtils.NamespaceInfo;
 }
-
-export const podStatusChildSource = {
-    async children(kubectl: Kubectl, parent: ResourceNode): Promise<ClusterExplorerNode[]> {
-        const nsarg = parent.namespace ? `--namespace=${parent.namespace}` : '';  // TODO: still not working!
-        const result = await kubectl.asJson<Pod>(`get pods ${parent.name} ${nsarg} -o json`);
-        if (result.succeeded) {
-            const pod = result.result;
-            let ready = 0;
-            pod.status.containerStatuses.forEach((status) => {
-                if (status.ready) {
-                    ready++;
-                }
-            });
-            return [
-                new MessageNode(`${pod.status.phase} (${ready}/${pod.status.containerStatuses.length})`),
-                new MessageNode(pod.status.podIP),
-            ];
-        }
-        else {
-            return [new MessageNode("Error", result.error[0])];
-        }
-    }
-};
-
-export const configItemsChildSource = {
-    async children(_kubectl: Kubectl, parent: ResourceNode): Promise<ClusterExplorerNode[]> {
-        const configData = parent.extraInfo!.configData;  // TODO: unbang
-        if (!configData || configData.length === 0) {
-            return [];
-        }
-        const files = Object.keys(configData);
-        return files.map((f) => new ConfigurationValueNode(configData, f, parent.kind, parent.name));
-    }
-};
-
-export const selectedPodsChildSource = {
-    async children(kubectl: Kubectl, parent: ResourceNode): Promise<ClusterExplorerNode[]> {
-        const labelSelector = parent.extraInfo!.labelSelector;  // TODO: unbang
-        if (!labelSelector) {
-            return [];
-        }
-        const pods = await kubectlUtils.getPods(kubectl, labelSelector);
-        return pods.map((p) => resourceNodeCreate(kuberesources.allKinds.pod, p.name, p.metadata, { podInfo: p }));
-    }
-};
-
-export const nodePodsChildSource = {
-    async children(kubectl: Kubectl, parent: ResourceNode): Promise<ClusterExplorerNode[]> {
-        const pods = await kubectlUtils.getPods(kubectl, null, 'all');
-        const filteredPods = pods.filter((p) => `node/${p.nodeName}` === parent.kindName);
-        return filteredPods.map((p) => resourceNodeCreate(kuberesources.allKinds.pod, p.name, p.metadata, { podInfo: p }));
-    }
-};
 
 export class SimpleResourceNode extends ResourceNode {
     constructor(kind: kuberesources.ResourceKind, name: string, metadata: ObjectMeta | undefined, extraInfo: ResourceExtraInfo | undefined) {
